Cleanup details pages
